# docs/swagger.yaml
definitions:
  model.CreateProposalResult:
    properties:
      proposal:
        type: string
      txid:
        type: string
    type: object
  model.ProposalResponseResult:
    properties:
      payload:
        description: payload to be signed as signedTx
        type: string
      proposal_responses:
        items:
          type: string
        type: array
    type: object
  model.TxResult:
    properties:
      info:
        type: string
      status:
        type: integer
    type: object
info:
  contact:
    email: david-khala@hotmail.com
  title: github.com/davidkhala/fabric-server-go
  version: v0.0.0
paths:
  /ecosystem/create-token:
    post:
      parameters:
      - description: Token owner
        in: formData
        name: owner
        required: true
        type: string
      - description: Token Content
        in: formData
        name: content
        required: true
        type: string
      - description: signer creator in bytes
        in: formData
        name: creator
        required: true
        type: string
      - description: Fabric channel name
        in: formData
        name: channel
        required: true
        type: string
      responses: {}
  /fabric/create-proposal:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Hex-encoded creator bytes
        in: formData
        name: creator
        required: true
        type: string
      - description: Fabric channel name
        in: formData
        name: channel
        required: true
        type: string
      - description: Fabric chaincode name
        in: formData
        name: chaincode
        required: true
        type: string
      - description: Fabric chaincode calling args, string array as JSON
        in: formData
        name: args
        required: true
        type: string
      - description: JSON format, like map[string]string
        in: formData
        name: transient
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateProposalResult'
  /fabric/ping:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: endpoint like grpc(s)://\<fqdn\> or \<fqdn\>
        in: formData
        name: address
        required: true
        type: string
      - description: Certificate in PEM format. should be in hex format after translation
          to solve linebreak issue
        in: formData
        name: certificate
        required: true
        type: string
      - description: pseudo endpoint \<fqdn\>
        in: formData
        name: ssl-target-name-override
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
  /fabric/transact/commit:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: json data to specify orderer
        in: formData
        name: orderer
        required: true
        type: string
      - description: serialized signed proposalResponses as envelop protobuf with
          hex format
        in: formData
        name: transaction
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TxResult'
  /fabric/transact/process-proposal:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: json data to specify endorsers
        in: formData
        name: endorsers
        required: true
        type: string
      - description: Hex-encoded and serialized signed-proposal protobuf
        in: formData
        name: signed-proposal
        required: true
        type: string
      - description: Hex-encoded and serialized proposal protobuf
        in: formData
        name: proposal
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProposalResponseResult'
  /ping:
    get:
      produces:
      - text/plain
      responses:
        "200":
          description: pong
          schema:
            type: string
swagger: "2.0"